rename the solution to TicketingHub
