--- src/API/ModularMonolith/ModularMonolith.http
@@ -1,6 +1,0 @@
-@ModularMonolith_HostAddress = http://localhost:5248
-
-GET {{ModularMonolith_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
